toradexhome/Acquisition/requirements.txt
python-can>=4.0.0
smbus2
pyserial
requests
